+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -985,9 +985,11 @@ class RequestListAPI(AuthorizationHandler):
         files = self.request.files
 
         for _file in files:
+            file_name = files[_file][0]["filename"]
             file_contents = files[_file][0]["body"]
 
             file_parameters[_file] = {
+                "filename": file_name,
                 "type": "bytes",
                 "base64": base64.b64encode(gzip.compress(file_contents)).decode(
                     "ascii"