+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -987,7 +987,7 @@ class RequestListAPI(AuthorizationHandler):
             file_contents = files[_file][0]["body"]
 
             file_parameters[_file] = {
-                "filename": file_name,
+                "file_name": file_name,
                 "type": "bytes",
                 "base64": base64.b64encode(gzip.compress(file_contents)).decode(
                     "ascii"