+++ src/app/beer_garden/db/mongo/models.py
@@ -189,7 +189,7 @@ class Parameter(MongoModel, EmbeddedDocument):
     display_name = StringField(required=False)
     optional = BooleanField(required=True, default=True)
     default = DynamicField(required=False, default=None)
-    filename = StringField(required=False)
+    file_name = StringField(required=False)
     description = StringField(required=False)
     choices = EmbeddedDocumentField("Choices", default=None)
     nullable = BooleanField(required=False, default=False)