+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TBD
 - Updated Pypi release package name from `beer-garden` to `beer_garden` to comply with PEP 625. 
 - Migrated from 408 to 504 HTTP error code for client defined wait time exceed. Required to prevent
   client from reattempting to submit the API request. 
+- Updated Authorization Filtering to return Downstream configuration info for only Garden Admins
 
 # 3.29.0